node track and utils to own files
