+++ src/metadata.ts
@@ -1,0 +1,1 @@
+export const metadata = new WeakMap();
